move connection handling code to ConnectionHandling
mv default_role to connection_handler
mv current_role to connection handling
mv default_connection_handler to connection_handling
mv shard_selector to connection_handling
mv connection_handler to connection handling
mv current_shard to connection_handling
mv current_connection to connection handling
mv connection_class methods to connection handling
mv default_shard to connection handling

Why:

ConnectionHandling was talking to methods in Core, even though there is
no explicit dependency. The code in question was all related to
connections and also not used elsewhere in Core, so it makes sense to
move that code into ConnectionHandling.
