+++ activerecord/lib/active_record/connection_adapters/abstract/connection_handler.rb
@@ -73,9 +73,12 @@ module ActiveRecord
         end
       end
 
+      attr_accessor :default_role
+
       def initialize
         # These caches are keyed by pool_config.connection_name (PoolConfig#connection_name).
         @connection_name_to_pool_manager = Concurrent::Map.new(initial_capacity: 2)
+        @default_role = ActiveRecord.writing_role
       end
 
       def prevent_writes # :nodoc:
